moved findBaseColor to widgethelper
