+++ src/util/widgethelper.h
@@ -49,6 +49,12 @@ inline QScreen* getScreen(
 #endif
 }
 
+// Get the base color of a widget, or recursively search the parent tree for one.
+//
+// Returns QColor(0,0,0) when none is found. As the recursion can go quite deep,
+// avoid calling repeatedly, but only when needed, e.g in a showEvent.
+QColor findBaseColor(QWidget* pWidget);
+
 } // namespace widgethelper
 
 } // namespace mixxx
